Apply checkstyle and pass all cases
